+++ test/cert/activation_pretest.h
@@ -1,0 +1,9 @@
+#ifndef cert_activation_pretest_h
+#define cert_activation_pretest_h
+
+
+#include "packet_util.h"
+
+void sp1_intial_join(TestServerState &server_state);
+
+#endif
